# Conflicts:
#	k9mail/build.gradle
#	k9mail/src/main/AndroidManifest.xml
#	k9mail/src/main/java/com/fsck/k9/controller/MessagingController.java
#	k9mail/src/main/java/com/fsck/k9/helper/MergeCursor.java
#	k9mail/src/main/res/drawable-hdpi/ic_notify_new_mail.png
#	k9mail/src/main/res/drawable-mdpi/ic_notify_new_mail.png
#	k9mail/src/main/res/drawable-xhdpi/ic_notify_new_mail.png
#	k9mail/src/main/res/values/constants.xml
#	k9mail/src/main/res/values/strings.xml

+++ k9mail/src/test/java/com/fsck/k9/helper/HtmlConverterTest.java
@@ -16,7 +16,7 @@ import static junit.framework.Assert.assertEquals;
 
 
 @RunWith(RobolectricTestRunner.class)
-@Config(manifest = Config.NONE)
+@Config(manifest = Config.NONE, sdk = 21)
 public class HtmlConverterTest {
     // Useful if you want to write stuff to a file for debugging in a browser.
     private static final boolean WRITE_TO_FILE = Boolean.parseBoolean(System.getProperty("k9.htmlConverterTest.writeToFile", "false"));
@@ -142,10 +142,10 @@ public class HtmlConverterTest {
         FileWriter fstream = null;
 
         try {
-            System.err.println(content);
-
             File f = new File(OUTPUT_FILE);
-            f.delete();
+            if (f.exists() && !f.delete()) {
+                throw new RuntimeException("Unable to delete existing output");
+            }
 
             fstream = new FileWriter(OUTPUT_FILE);
             BufferedWriter out = new BufferedWriter(fstream);
